add typescript support convert extentions add pretty script in package
