+++ src/App.css
@@ -1,6 +1,6 @@
 .App {
-    text-align: center;
-    font-size: 1.5rem;
-    max-width:50rem;
-    margin: 2rem auto;
+  text-align: center;
+  font-size: 1.5rem;
+  max-width: 50rem;
+  margin: 2rem auto;
 }
