+++ src/components/Button/Button.css
@@ -1,27 +1,27 @@
 .button {
-    border: 1px solid #ccc;
-    background-color: #ddd;
-    border-radius: 5px;
-    font-family: inherit;
-    font-size:1.2rem;
-    line-height:1;
-    padding: .75rem 1.25rem;
-    min-width:10rem;
-    margin: auto .5rem;
-    font-weight: bold;
-    letter-spacing: 1px;
+  border: 1px solid #ccc;
+  background-color: #ddd;
+  border-radius: 5px;
+  font-family: inherit;
+  font-size: 1.2rem;
+  line-height: 1;
+  padding: 0.75rem 1.25rem;
+  min-width: 10rem;
+  margin: auto 0.5rem;
+  font-weight: bold;
+  letter-spacing: 1px;
 }
 
 .button-primary {
-    background-color:rebeccapurple;
-    color: #efefef;
+  background-color: rebeccapurple;
+  color: #efefef;
 }
 
 .button-secondary {
-    background-color:mediumturquoise;
-    color: #efefef;
+  background-color: mediumturquoise;
+  color: #efefef;
 }
 .button-tertiary {
-    border-color: #1246f1;
-    background-color: transparent;
+  border-color: #1246f1;
+  background-color: transparent;
 }
