--- src/components/Button/index.jsx
@@ -1,1 +1,0 @@
-export * from './Button';
